+++ src/models/classes/Translation.php
@@ -203,8 +203,14 @@ class Translation
                 'submit' => $this->l('modal.premium.submit', 'translation'),
             ],
             'applepay' => [
-                'text' => $this->l('modal.applepay.text', 'translation'),
-                'submit' => $this->l('modal.applepay.submit', 'translation'),
+                'display' => [
+                    'text' => $this->l('modal.applepay.display.text', 'translation'),
+                    'submit' => $this->l('modal.applepay.display.submit', 'translation'),
+                ],
+                'carrier' => [
+                    'text' => $this->l('modal.applepay.carrier.text', 'translation'),
+                    'submit' => $this->l('modal.applepay.carrier.submit', 'translation'),
+                ],
             ],
             'error' => [
                 'text' => $this->l('modal.error.text', 'translation'),
@@ -508,17 +514,17 @@ class Translation
                 ],
                 'link' => $this->l('paymentmethods.applepay.link', 'translation'),
                 'call_to_action' => $this->l('paymentmethods.applepay.call_to_action', 'translation'),
-                'cart' => [
-                    'title' => $this->l('paymentmethods.applepay.cart.title', 'translation'),
-                    'description' => $this->l('paymentmethods.applepay.cart.description', 'translation'),
-                ],
-                'checkout' => [
-                    'title' => $this->l('paymentmethods.applepay.checkout.title', 'translation'),
+                'display' => [
+                    'title' => $this->l('paymentmethods.applepay.display.title', 'translation'),
+                    'cart' => $this->l('paymentmethods.applepay.display.cart', 'translation'),
+                    'checkout' => $this->l('paymentmethods.applepay.display.checkout', 'translation'),
+                    'product' => $this->l('paymentmethods.applepay.display.product', 'translation'),
                 ],
                 'carrier' => [
                     'title' => $this->l('paymentmethods.applepay.carrier.title', 'translation'),
                     'alert' => $this->l('paymentmethods.applepay.carrier.alert', 'translation'),
                     'description' => $this->l('paymentmethods.applepay.carrier.description', 'translation'),
+                    'instructions' => $this->l('paymentmethods.applepay.carrier.instructions', 'translation'),
                 ],
                 'modal' => [
                     'subtotal' => $this->l('paymentmethods.applepay.modal.subtotal', 'translation'),