+++ tests/actions/ConfigurationAction/saveActionTest.php
@@ -89,12 +89,15 @@ class saveActionTest extends BaseConfigurationAction
         );
     }
 
-    public function testWhenApplepayCarriersIsEmpty()
+    public function testWhenNoApplepayDisplayIsSelected()
     {
         $datas = new \stdClass();
         $datas->action = 'payplug_save_data';
         $datas->enable_applepay = true;
-        $datas->enable_applepay_cart = true;
+        $datas->payplug_applepay_display = new \stdClass();
+        $datas->payplug_applepay_display->cart = false;
+        $datas->payplug_applepay_display->checkout = false;
+        $datas->payplug_applepay_display->product = false;
         $datas->applepay_carriers = [];
 
         $this->assertSame(
@@ -102,8 +105,8 @@ class saveActionTest extends BaseConfigurationAction
                 'success' => false,
                 'data' => [
                     'title' => null,
-                    'msg' => 'modal.applepay.text',
-                    'close' => 'modal.applepay.submit',
+                    'msg' => 'modal.applepay.display.text',
+                    'close' => 'modal.applepay.display.submit',
                     'class' => '-error',
                 ],
             ],
@@ -111,12 +114,40 @@ class saveActionTest extends BaseConfigurationAction
         );
     }
 
+    public function testWhenApplepayCarriersIsEmpty()
+    {
+        $datas = new \stdClass();
+        $datas->action = 'payplug_save_data';
+        $datas->enable_applepay = true;
+        $datas->payplug_applepay_display = new \stdClass();
+        $datas->payplug_applepay_display->cart = true;
+        $datas->applepay_carriers = [];
+
+        $this->assertSame(
+            [
+                'success' => false,
+                'data' => [
+                    'title' => null,
+                    'msg' => 'modal.applepay.carrier.text',
+                    'close' => 'modal.applepay.carrier.submit',
+                    'class' => '-error',
+                ],
+            ],
+            $this->action->saveAction($datas)
+        );
+    }
+
+    /**
+     * @group debug
+     */
     public function testWhenConfigurationCannotBeUpdate()
     {
         $datas = new \stdClass();
         $datas->action = 'payplug_save_data';
+        $datas->enable_applepay = true;
         $datas->applepay_carriers = ['1', '2', '4'];
-        $datas->enable_applepay_cart = true;
+        $datas->payplug_applepay_display = new \stdClass();
+        $datas->payplug_applepay_display->cart = true;
         $datas->enable_standard = 1;
 
         $this->configuration->shouldReceive([
@@ -127,7 +158,7 @@ class saveActionTest extends BaseConfigurationAction
             [
                 'success' => false,
                 'data' => [
-                    'message' => 'An error has occurred while register payplug_applepay_carriers',
+                    'message' => 'An error has occurred while register applepay_carriers',
                 ],
             ],
             $this->action->saveAction($datas)
@@ -138,8 +169,10 @@ class saveActionTest extends BaseConfigurationAction
     {
         $datas = new \stdClass();
         $datas->action = 'payplug_save_data';
+        $datas->enable_applepay = true;
         $datas->applepay_carriers = ['1', '2', '4'];
-        $datas->enable_applepay_cart = true;
+        $datas->payplug_applepay_display = new \stdClass();
+        $datas->payplug_applepay_display->cart = true;
         $datas->payplug_standard = 1;
 
         $this->configuration->shouldReceive([